requirements.txt
google-generativeai
InquirerPy
pyfiglet
colorama
python-dotenv
terminal-text-effects